Fix trivial -Xlint:all and Sonar findings
-Xlint:all findings (raw type, missing-explicit-ctor), both from a
one-off diagnostic sweep, not a build-enforced check:
- FileStorePanel: suppress rawtypes alongside the existing unchecked
  suppression on the same generic-array-creation idiom.
- SystemInfoTest (FFM): add an explicit package-private constructor
  so the exported oshi.ffm package doesn't expose a default public
  one.

Also widen oshi-core-ffm's org.slf4j requirement to transitive:
ForeignFunctions' public callInArena*OrDefault methods take a Logger
parameter directly, unlike other modules where Logger is only a
private field, so this one actually needs it.

Sonar findings from the latest master analysis:
- S108: comment the intentionally-empty default arm in
  LogicalProcessorInformationFFM, matching the JNA twin's existing
  "Ignore Group info" comment.
- S1905: drop the genuinely-redundant int-to-long cast in
  KstatUtilFFM (the JNA twin already omits it). The three long-to-
  double casts Sonar also flagged in FileStorePanel/MemoryPanel are
  NOT redundant - they're required by Error Prone's LongDoubleConversion
  check (ERROR severity as of #3606), which exists specifically to
  catch long-to-double precision loss. Suppressed those three with
  NOSONAR instead of removing them.

Left alone: S9149 (22 findings) - MSFTStorageFFM/JNA and similar
classes redeclare identical static methods from their oshi-common
base class purely so each backend module can reference a locally-
named class. Real callers exist for all of them; fixing this means
deleting ~9 shim class-pairs and repointing callers at the shared
base classes directly across both twins - a separate, larger PR.

+++ oshi-core-ffm/src/main/java/oshi/ffm/util/platform/unix/solaris/KstatUtilFFM.java
@@ -256,7 +256,7 @@ public final class KstatUtilFFM {
             MemorySegment named = data.reinterpret(KSTAT_NAMED_LAYOUT.byteSize());
             byte dt = namedDataType(named);
             return switch (dt) {
-                case KSTAT_DATA_INT32 -> (long) namedValueInt32(named);
+                case KSTAT_DATA_INT32 -> namedValueInt32(named);
                 case KSTAT_DATA_UINT32 -> FormatUtil.getUnsignedInt(namedValueInt32(named));
                 case KSTAT_DATA_INT64, KSTAT_DATA_UINT64 -> namedValueInt64(named);
                 default -> {
